% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_uniprot_data.R
\name{create_uniprot_data}
\alias{create_uniprot_data}
\title{Download aaaView UniProt (Swiss Pro) data}
\usage{
create_uniprot_data(local = FALSE, uniprot_sprot = "uniprot_sprot.fasta.gz")
}
\value{
Puts peviz_uniprot_data.Rdata in app folder
}
\description{
Download aaaView UniProt (Swiss Pro) data
}
\examples{
\dontrun{create_uniprot_data()}
}
