% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaView.R
\name{launch}
\alias{launch}
\title{Wrapper that starts the aaaView Shiny App.}
\usage{
launch(app_path = system.file("app", package = "aaaView"))
}
\arguments{
\item{app_path}{The path to the aaaView Rlibrary folder holding the
server.R and ui.R files.}
}
\value{
Loads shiny app.
}
\description{
Wrapper that starts the aaaView Shiny App.
}
