% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_msa.R
\name{tidy_msa}
\alias{tidy_msa}
\title{tidy_msa from Guangchuang Yu's ggmsa}
\usage{
tidy_msa(msa, start = NULL, end = NULL)
}
\arguments{
\item{msa}{multiple sequence alignment file or sequence object in
DNAStringSet, RNAStringSet, AAStringSet, BStringSet, DNAMultipleAlignment,
RNAMultipleAlignment, AAMultipleAlignment, DNAbin or AAbin}

\item{start}{start position to extract subset of alignment}

\item{end}{end position to extract subset of alignemnt}
}
\value{
tibble data frame
}
\description{
Convert msa file/object to tidy data frame.
}
\author{
Guangchuang Yu
}
